fix: increase init delay for liveness probe
The initial delay in the liveness probe config for k8s health checks was
too short. This was due to the mistaken assumption that k8s would only
start running the liveness probe after the readiness probe succeeded.

For services that take a long time to start up, the old config would
sometimes lead to the liveness probe failing and thus the container
being restarted - repeatedly, in some situations.

The initial delay and frequency of the readiness probe were also
reduced.

Later, we should aim to make initialDelaySeconds configurable for the
liveness probe.