refactor: rename task baseKey to key and key to id
In short, baseKey is now called key, key is now called id, and id is now
called uid. The same applies to the corresponding getter methods on
BaseTask.

baseKey has always been a somewhat awkward term, and with some of the
recent changes to the events emitted by the TaskGraph, it had started to
become slightly confusing as well. The new terminology should be clearer
and more intuitive.